Create types.ts file
